media_codecs: fix Netflix
The current Netflix app tries very hard to find a codec which supports
AdaptivePlayback. As our HW-codec doesn't support that, the Google
H.264 software-codec gets used. But it is incapable of decoding the
stream from Netflix, returning H264SWDEC_STRM_ERR. The hw-codec
plays the stream just fine.
This removes the declaration of adaptive-playback from the (broken)
software-codec which causes Netflix to use hw-based decoding.