Allow display blanking during boot sequence
This is a short init.d script that listens for button input
and toggles blanking of the LCD panel. This should be useful
for dexopt boot sequences, which result in lengthy cache
rebuilding process, and the dreaded "Android is Upgrading"
dialog.

It responds to either the power button or the volume buttons.

I experimented with the option of building this directly into
android.policy, but opted for this init.d script instead so that
the device is never actually put to sleep, and the cache building
process is allowed to proceed using full awake CPU power.

This script was written for Tuna, but can be adapted to work with
other devices by substituting the correct display and event parameters.

(For Tuna, the buttons are input/event2 and the display is
omapdss/display0).