Implemented SELinux rules for tuna.
Note: the implementation is quite hacky because of the tunasetup.sh
script. The set of SELinux Android policies is not really designed
for a case when some script remounts /system to rw at runtime and
moves the files around - most of these actions are explicitly
forbidden by 'neverallow' rules. Running these actions in 'recovery'
domain, while hacky, seems to be the cleanest solution among all
the alternatives.